post request to server works and updates state
